add results

+++ src/Layout.hpp
@@ -33,4 +33,4 @@ struct Layout {
 };
 
 Layout create_layout(const std::vector<Container>& containers);
-float cost(const Layout& layout, const std::vector<Container>& containers);
+int cost(const Layout& layout, const std::vector<Container>& containers);
